move and fix waffle tests
